refactor: extract wheel demo into reusable component
